Update /etc/init
If filesystem checks fail, continue with a read-only root filesystem
so the machine can be repaired over ssh. In this case, don't attempt to
start user init scripts.

Run ntpd with daemon -w /run/resolv.conf (to wait for iwd to find the
network) rather than daemon await-net now the CLOCK_MONOTONIC bug is
fixed so it starts up reliably.

Wait for /dev/log to exist with daemon -w before launching iwd. It will
fail to log at all if it wins the race against syslogd.